requirements.txt
telethon
python-dotenv
requests
pybit
